one time alarm done and refactoring
